Change getMethod function
I changed the format that the response data of the GET method was
being received. It is now json format instead of text. This makes
searching for parameters much more easier, with the reponse now
being a object with paramers instead of a long string that requires
RegEx to be used!