Create setProperties functions
The way that regular variables and pointers are found in the code
of the user have changed. It now uses the matchAll function
inside the setProperties function for finding and setting
properties for both of them. With matchAll function capturing
groups in RegEx can be used to identify properties of variables
in a much more efficient way. That is why the function getNames
and getTypes were deleted. The function getContents was modified
for working with one object at the time.