Complete redesign of the site
I have changed positions and appearance of elements, added the p5
library, used p5 functions to draw the memory representation in a
canvas and changed the site's name to "Como Ponteiros Funcionam"
(How Pointers work). The representation of the memory is temporary
as it has lots of bugs, but it is a good prototype to demonstrate
what I want to achieve.