Change method of key recognition
Instead of using the keycodes that changes from browser to browse,
I analyze the last character typed in the textbox looking for
where the text cursor (caret) is. This way I can know if the user
typed "enter" or a semicolon regardless of the browser they are
using.
A bug that happened when the user declared a pointer without
initializing it was also fixed.